feat(dispositions): allow float values for timeout minutes
Update the disposition model and schemas to support floating-point
numbers for the `timeout_minutes` field. This allows for more precise
timeout configurations.

- Change `timeout_minutes` type from `Integer` to `Float` in SQLAlchemy model
- Update Pydantic schemas to use `float` instead of `int`
- Adjust validation constraints from `ge=1` to `gt=0`

+++ backend/app/models/disposition.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime, timezone
-from sqlalchemy import Column, String, Boolean, Integer, DateTime
+from sqlalchemy import Column, String, Boolean, Integer, Float, DateTime
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from app.core.database import Base
@@ -12,7 +12,7 @@ class Disposition(Base):
     name = Column(String(255), nullable=False)
     category = Column(String(100), nullable=False, default="Negociação")
     has_timeout = Column(Boolean, default=True, nullable=False)
-    timeout_minutes = Column(Integer, nullable=True)
+    timeout_minutes = Column(Float, nullable=True)
     is_active = Column(Boolean, default=True, nullable=False)
     created_at = Column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc), nullable=False)
